+++ Package.swift
@@ -16,8 +16,8 @@ let package = Package(
             targets: ["KeyboardKit"])
     ],
     dependencies: [
-        .package(url: "https://github.com/Quick/Quick.git", .upToNextMajor(from: "4.0.0")),
-        .package(url: "https://github.com/Quick/Nimble.git", .upToNextMajor(from: "9.0.0")),
+        .package(url: "https://github.com/danielsaidi/Quick.git", .branch("main")), // .upToNextMajor(from: "4.0.0")),
+        .package(url: "https://github.com/danielsaidi/Nimble.git", .branch("main")), // .upToNextMajor(from: "9.0.0")),
         .package(url: "https://github.com/danielsaidi/MockingKit.git", .upToNextMajor(from: "1.0.0"))
     ],
     targets: [